Новый алгоритм распознования..
1. Добавлена генерация быстрых ответов для оптимизации запросов
2. Исправлена ошибка окна смены пути если приложение не имеет его вовсе
3. Изменен алгоритм распознования голоса STT, упрощены обработки команд и запуск алгоритма ассистента
4. Небольшие правки

+++ RVC__main.py
@@ -129,6 +129,8 @@ class generateTTS():
         speed, tts_voice,
         f0_up_key, f0_method,
         index_rate, protect,
+        pattern = False,
+        filename = None,
         filter_radius=3,
         resample_sr=0,
         rms_mix_rate=0.25
@@ -200,8 +202,12 @@ class generateTTS():
             )
             if self.tgt_sr != resample_sr >= 16000: self.tgt_sr = resample_sr
 
-            filename = Pathlib_y.get_mainLOCALpath()+"/result.wav"
-            os.remove(filename)
+            if pattern:
+                filename = Pathlib_y.get_voicePatternspath()+"/"+filename
+            else:
+                filename = Pathlib_y.get_mainLOCALpath()+"/result.wav"
+                os.remove(filename)
+            
             processing_utils.audio_to_file(self.tgt_sr, audio_opt, filename, 'wav')
 
             print(f"tts-out || Success. Time: edge-tts: {edge_time}s, npy: {times[0]}s, f0: {times[1]}s, infer: {times[2]}s")
